Moved contact page out to its own app
