+++ shellyeah/shellyeah/settings.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'lottery_sim',
     'players',
+    'contact',
     'bootstrap5'
 ]
