+++ shellyeah/shellyeah/urls.py
@@ -20,5 +20,6 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('lottery_sim/', include('lottery_sim.urls') ),
+    path('contact/', include('contact.urls') ),
     path('',include('players.urls')),
 ]
